requirements.txt
google-cloud-bigquery
joblib
kafka-python
pandas
sentence-transformers
sklearn

-e .
